+++ lib/maze/hooks/browser_hooks.rb
@@ -8,29 +8,39 @@ module Maze
         case config.farm
         when :bb
           if ENV['BUILDKITE']
-            credentials = Maze::BitBarUtils.account_credentials config.tms_uri
+            credentials = Maze::Client::BitBarClientUtils.account_credentials config.tms_uri
             config.username = credentials[:username]
             config.access_key = credentials[:access_key]
           end
           tunnel_id = SecureRandom.uuid
-          config.capabilities = Maze::Capabilities.for_bitbar_browsers config.browser,
-                                                                       config.access_key,
-                                                                       tunnel_id,
-                                                                       config.capabilities_option
+          capabilities = Selenium::WebDriver::Remote::Capabilities.new
+          capabilities['bitbar_apiKey'] = config.access_key
+          browsers = YAML.safe_load(File.read("#{__dir__}/../client/selenium/bb_browsers.yml"))
+          capabilities.merge! browsers[config.browser]
+          capabilities.merge! JSON.parse(config.capabilities_option)
+          config.capabilities = capabilities
 
-          Maze::SmartBearUtils.start_local_tunnel config.sb_local,
-                                                  config.username,
-                                                  config.access_key
+          Maze::Client::BitBarClientUtils.start_local_tunnel config.sb_local,
+                                                             config.username,
+                                                             config.access_key
           tunnel_id
         when :bs
           # BrowserStack browser
           tunnel_id = SecureRandom.uuid
-          config.capabilities = Maze::Farm::BrowserStack::Capabilities.browser config.browser,
-                                                                               tunnel_id,
-                                                                               config.capabilities_option
-          Maze::Farm::BrowserStack::Utils.start_local_tunnel config.bs_local,
-                                                             tunnel_id,
-                                                             config.access_key
+          capabilities = {
+            'bstack:options' => {
+              'local' => 'true',
+              'localIdentifier' => tunnel_id
+            }
+          }
+          browsers = YAML.safe_load(File.read("#{__dir__}/../client/selenium/bs_browsers.yml"))
+          capabilities.deep_merge! browsers[config.browser]
+          capabilities.deep_merge! JSON.parse(config.capabilities_option)
+          config.capabilities = Selenium::WebDriver::Remote::Capabilities.new capabilities
+
+          Maze::Client::BrowserStackClientUtils.start_local_tunnel config.bs_local,
+                                                                   tunnel_id,
+                                                                   config.access_key
         end
 
         # Create and start the relevant driver
@@ -53,7 +63,7 @@ module Maze
 
       def at_exit
         if Maze.config.farm == :bs
-          Maze::Farm::BrowserStack::Utils.stop_local_tunnel
+          Maze::Client::BrowserStackClientUtils.stop_local_tunnel
           Maze.driver.driver_quit
         end
       end