+++ CHANGELOG.md
@@ -18,7 +18,9 @@
 
 ## Refactor
 
-- Refactor BrowserStack specific Appium/Selenium code into `Maze::Farm::BrowserStack` [390](https://github.com/bugsnag/maze-runner/pull/390)
+- Refactor Appium/Selenium client code into `Maze::Client`
+  [390](https://github.com/bugsnag/maze-runner/pull/390)
+  [399](https://github.com/bugsnag/maze-runner/pull/399)
 
 # 7.0.0 - 2022/07/29